Fix ArticleExtractor not in scope for extension targets
ArticleExtractor is excluded from the Add Feed and Widgets targets,
so RSSParser+HTML.swift cannot reference it directly. Extract the
ad-label patterns into a standalone AdvertisementTextFilter in
Shared/Domain Options/ (available to all targets) and have both
ArticleExtractor and RSSParser delegate to it.
