+++ Shared/Article Extractor/ArticleExtractor+Noise.swift
@@ -532,20 +532,9 @@ extension ArticleExtractor {
         }
     }
 
-    /// Standalone ad-label text patterns typically injected by news sites
-    /// (e.g. NYTimes "ADVERTISEMENT" paragraphs).
-    static let advertisementTextPatterns: Set<String> = [
-        "advertisement",
-        "advertising",
-        "sponsored content",
-        "paid post",
-        "paid content"
-    ]
-
     /// Returns true when the trimmed, lowercased text matches a known ad label.
     static func isAdvertisementText(_ text: String) -> Bool {
-        let trimmed = text.trimmingCharacters(in: .whitespacesAndNewlines).lowercased()
-        return advertisementTextPatterns.contains(trimmed)
+        AdvertisementTextFilter.isAdvertisementText(text)
     }
 
     /// Removes block-level elements (`<p>`, `<div>`, `<span>`) whose only
